Added room groups
