+++ package.json
@@ -46,6 +46,9 @@
       "tsx",
       "js"
     ],
+    "moduleNameMapper": {
+      "\\.(s?css|less)$": "<rootDir>/src/jest/style.js"
+    },
     "testRegex": "(/__tests__/.*|\\.test)\\.(ts|tsx|js)$",
     "transform": {
       ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
