+++ src/flex/index.ts
@@ -1,0 +1,13 @@
+import './style.scss';
+
+export const flex = (value: number | string) => {
+  value = (value == null ? '' : value).toString();
+
+  return {
+    '-webkit-box-flex': value,
+    '-moz-box-flex': value,
+    '-webkit-flex': value,
+    '-ms-flex': value,
+    'flex': value
+  };
+};
