+++ src/model/index.ts
@@ -1,2 +1,11 @@
-import Model from './model';
-export default Model;
+export default class Model<T> {
+  constructor(...data: Partial<T>[]) {
+    data.forEach((item) => {
+      for (let key in item) {
+        if (item.hasOwnProperty(key)) {
+          (<any>this)[key] = (<any>item)[key];
+        }
+      }
+    });
+  }
+}
