+++ src/room-list/index.ts
@@ -1,9 +1,11 @@
 import './style.scss';
 
 import { div } from 'compote/html';
+import groupBy = require('lodash/groupBy');
+import keys = require('lodash/keys');
 
+import { flex } from '../flex';
 import { Room, selectRoom } from '../room';
-import store from '../store';
 
 export const RoomListItem = (selectedRoom: Room) => (room: Room) => (
   div({
@@ -12,9 +14,15 @@ export const RoomListItem = (selectedRoom: Room) => (room: Room) => (
   }, room.name)
 );
 
-export const RoomList = (rooms: Room[]) => {
-  const { selectedRoom } = store.getState();
-  return div({ className: 'room-list flex-row justify-content-start align-items-stretch' }, rooms.map(
-    RoomListItem(selectedRoom)
-  ));
+export const RoomList = (rooms: Room[], selectedRoom: Room) => {
+  const groupedRooms = groupBy(rooms, 'group');
+  const groups = keys(groupedRooms);
+  return div({ className: 'room-list flex-row-md justify-content-start align-items-stretch' }, groups.map((group) => (
+    div({ className: 'flex-item', style: flex(1) }, [
+      div({ className: 'room-list-group' }, group),
+      div({ className: 'flex-row justify-content-start align-items-stretch' },
+        groupedRooms[group].map(RoomListItem(selectedRoom))
+      )
+    ])
+  )));
 };
