+++ src/room/index.test.ts
@@ -1,7 +1,7 @@
 import 'jest';
 
 import Model from '../model';
-import { Room/*, getRooms, addRoom, selectRoom*/ } from './room';
+import { Room/*, getRooms, addRoom, selectRoom*/ } from './index';
 
 test('should extend Model class', () => {
   expect(new Room()).toBeInstanceOf(Model);
