+++ src/style.scss
@@ -2,7 +2,6 @@
 @import 'style/mixins';
 @import 'style/reset';
 @import 'style/global';
-@import 'style/classes';
 @import 'style/animations';
 
 // Add padding to root container element
