+++ src/supply/index.test.ts
@@ -1,7 +1,7 @@
 import 'jest';
 
 import Model from '../model';
-import { Supply/*, getSupplies, addSupply, requestSupply, supplyRequested*/ } from './supply';
+import { Supply/*, getSupplies, addSupply, requestSupply, supplyRequested*/ } from './index';
 
 test('should extend Model class', () => {
   expect(new Supply()).toBeInstanceOf(Model);
